+++ src/junker.ts
@@ -1,6 +1,5 @@
 // import util from 'util';
-import { JsonTokenTypes } from "./tokenize.js";
-import { JsonToken, ParseSettings } from "./types.js";
+import { JsonTokenTypes, JsonToken, ParseSettings } from "./types.js";
 
 function findLastTokenIndexIn(tokenList, JsonTokenTypes): number {
   let rindex = tokenList.length - 1;
