+++ src/types.ts
@@ -1,5 +1,22 @@
 import { IJsonNode } from "./ast.js";
 import { JsonPosition } from "./position.js";
+
+export enum JsonTokenTypes {
+  COMMENT = "COMMENT", // // ... \n\r? or /* ... */
+  LEFT_BRACE = "LEFT_BRACE", // {
+  RIGHT_BRACE = "RIGHT_BRACE", // }
+  LEFT_BRACKET = "LEFT_BRACKET", // [
+  RIGHT_BRACKET = "RIGHT_BRACKET", // ]
+  COLON = "COLON", //  :
+  COMMA = "COMMA", // ,
+  STRING = "STRING", //
+  NUMBER = "NUMBER", //
+  TRUE = "TRUE", // true
+  FALSE = "FALSE", // false
+  NULL = "NULL", // null
+  IDENTIFIER = "IDENTIFIER", // identifiers
+}
+
 export interface ParseResult<T extends IJsonNode> {
   value: T;
   index: number;
@@ -11,7 +28,7 @@ export interface ParseSettings {
 }
 
 export interface JsonToken {
-  type: any;
-  value: any;
+  type: JsonTokenTypes;
+  value: string;
   position?: JsonPosition;
 }
